+ writing for Hemera photo-objects and Heretic II textures, and a sampling overflow
The photo-object states where its picture begins rather than keeping it at a fixed offset, so the
writer states it too. A real one carries a second picture after the first — a palette-indexed PNG
none of the samples explains — and that is not written, because writing something whose meaning is
unknown is inventing it.

A Heretic texture has room for sixteen mipmap levels. Only the first is filled and the rest stay at
nought, which says the file holds no smaller copies rather than that it holds empty ones at offset
zero.

The sampling arithmetic was in int, so a source wider than about 32768 wrapped negative part way
along a row and threw. Several headers here state a width up to 65535, which made the widest picture
those formats allow the one that could not be sampled onto a screen.

+++ FileFormat.Core/RawImageExtensions.cs
@@ -144,10 +144,13 @@ public static class RawImageExtensions {
 
     var rgb = new byte[width * height * 3];
     for (var y = 0; y < height; ++y) {
-      var sourceY = y * image.Height / height;
+      // In long arithmetic. A source wider than about 32768 overflows a signed int part way along a
+      // row, and the offset comes out negative — so the widest pictures the headers here can state
+      // were the ones that threw.
+      var sourceY = (int)((long)y * image.Height / height);
 
       for (var x = 0; x < width; ++x) {
-        var from = (sourceY * image.Width + x * image.Width / width) * 3;
+        var from = (int)(((long)sourceY * image.Width + (long)x * image.Width / width) * 3);
         var to = (y * width + x) * 3;
         rgb[to] = source.PixelData[from];
         rgb[to + 1] = source.PixelData[from + 1];
